+++ src/yafowil/tests.py
@@ -18,6 +18,7 @@ TESTFILES = [
     'compound.rst',
     'controller.rst',
     'table.rst',
+    'resources.rst',
 ]